Fix marshalling/unmarshalling of enums in some situations
In the codebase we are working in, enums in ent are defined with
lower-case values, and replacing our marshal funcs with the ones
in ent-contrib for gqlgen caused errors.

This PR attempts to check the value lower-cased, if the value given
does not validate properly, and use (and validate) that value. It
also updates the marshalling to the SCREAMING_SNAKE_CASE that is
expected by GraphQL