collection.tmpl: spare 2nd query to db on nested pagination
When paginating nested fields, we first query the neighbors nodes and then
calls `loadTotal` func to calculate the total count of the nested field.

This change spares the query to the db that count node's neighbors
by checking if node has no neighbors and if so, update loadedTotal with zeros.