entgql: support SkipMutationCreateInput (and update) on entity
Before this change, when SkipMutationCreateInput was annotated on an
entity, the CreateXxxInput was omitted from the gql_mutation_input.go
generated types, but still included in the ent.graphql generated schema.
The same was true for SkipMutationUpdateInput.

To work around this, you could use entgql.Mutations to select only the
ones you want generated. But, according to the annotation comments,
using SkipMutationCreateInput in this way should be supported.

This change adds the proper checks to buildMutationInputs to omit the
specified mutation inputs from the GraphQL schema.