entproto: fix service.List pageToken for custom fields (UUID/string)
The current List service template generates faulty
code if a custom GoType field is used (currently only UUID and string
supported).

While not tested on a large set of inputs, using a call to the
`field_to_ent` template hopefully leads to a more sane behaviour,
as it is also used for Get() and Update() function calls.