entgql: pagination name overrides
When a GraphQL type extends a definition in another sub-graph the names
of the pagination types clash, in particular the where-inputs.  Rather
than taking the union of the properties when creating the supergraph, it
takes the common properties.  This is true whether your input types are
declared with `type` or `extend type` (or so it seems).  It doesn't make
sense to copy-paste the where-input types between projects, since Ent
cannot use them (would mean joining across databases).

This commit allows the rename of the pagination types, avoiding the
clash described above.  The feature is used as follows:

	entgql.PaginationNameOverrides = map[string]string{
		"Domain1_EntityA":  "DomainN_Domain1_EntityA",
		"Domain2_EntityA":  "DomainN_Domain2_EntityB",
	}

In this example, our sub-graph (DomainN) is providing type extensions
for entities in two different sub-graphs (Domain1, Domain2).  In this
manner, we will not get collisions when multiple sub-graphs extend a
particular type.

This is combined with a `entgql.WithOutputWriter` that uses a fork of
`github.com/vektah/gqlparser/v2@v2.5.20/formatter/formatter.go` that
overrides `FormatSchema` to accept a set of extension types.